根据老师意见修改

+++ src/main/java/com/controller/IndiceController.java
@@ -86,9 +86,11 @@ public class IndiceController {
 		response.setContentType("text/json;charset=utf-8");
 		response.getWriter().write(JSON.toJSONString(data));
 	}
+
 	/**
 	 * 获取一名用户的所有的算子
-	 * @param user_id 用户id
+	 * 
+	 * @param user_id  用户id
 	 * @param response
 	 * @throws IOException
 	 */
@@ -99,9 +101,11 @@ public class IndiceController {
 		response.setContentType("text/json;charset=utf-8");
 		response.getWriter().write(JSON.toJSONString(data));
 	}
+
 	/**
 	 * 获取一名用户没有的算子
-	 * @param user_id 用户id
+	 * 
+	 * @param user_id  用户id
 	 * @param response
 	 * @throws IOException
 	 */
@@ -112,38 +116,55 @@ public class IndiceController {
 		response.setContentType("text/json;charset=utf-8");
 		response.getWriter().write(JSON.toJSONString(data));
 	}
+
 	/**
 	 * 为用户增添算子
-	 * @param user_id 用户id
+	 * 
+	 * @param user_id        用户id
 	 * @param selectedAddOps 要增添的算子的id
 	 * @param response
 	 * @throws IOException
 	 */
 	@RequestMapping(params = "request=addUserOperators")
-	public void addUserOperators(String user_id, String selectedAddOps, HttpServletResponse response) throws IOException {
+	public void addUserOperators(String user_id, String selectedAddOps, HttpServletResponse response)
+			throws IOException {
 		System.out.println("捕获到参数为request=addUserOperators的url请求");
 		int id = Integer.parseInt(user_id);
-		List<Integer> list = JSONObject.parseArray(selectedAddOps, Integer.class);//解析出要删除的用户的算子
+		List<Integer> list = JSONObject.parseArray(selectedAddOps, Integer.class);// 解析出要删除的用户的算子
 		int num = this.indiceDao.addUserOperators(id, list);
 		response.setContentType("text/json;charset=utf-8");
 		response.getWriter().write(JSON.toJSONString(num));
 	}
+
 	/**
 	 * 删除用户的算子
-	 * @param user_id 用户id
+	 * 
+	 * @param user_id        用户id
 	 * @param selectedDelOps 用户选中的要删除算子
 	 * @param response
 	 * @throws IOException
 	 */
 	@RequestMapping(params = "request=deleteUserOperators")
-	public void deleteUserOperators(String user_id, String selectedDelOps, HttpServletResponse response) throws IOException {
+	public void deleteUserOperators(String user_id, String selectedDelOps, HttpServletResponse response)
+			throws IOException {
 		System.out.println("捕获到参数为request=deleteUserOperators的url请求");
 		int id = Integer.parseInt(user_id);
-		List<OperatorInfo> list = JSONObject.parseArray(selectedDelOps, OperatorInfo.class);//解析出要删除的用户的算子
-		int num = this.indiceDao.deleteUserOperators(id, list);
+		List<OperatorInfo> list = JSONObject.parseArray(selectedDelOps, OperatorInfo.class);// 解析出要删除的用户的算子
+		List<Integer> existedList = this.indiceDao.selectUsedOperators(id);
+		int num = 0;
+		for (OperatorInfo op : list) {
+			if (existedList.contains(op.getOperator_id())) {
+				num = -1;
+				break;
+			}
+		}
+		if (num != -1) {
+			num = this.indiceDao.deleteUserOperators(id, list);
+		}
 		response.setContentType("text/json;charset=utf-8");
 		response.getWriter().write(JSON.toJSONString(num));
 	}
+
 	/**
 	 * 退出登录
 	 * 
@@ -281,7 +302,7 @@ public class IndiceController {
 		SchemeInfo scheme = JSON.parseObject(line, SchemeInfo.class);// 把JSON字符串解析为JavaBean
 		List<SchemeInfo> list = this.indiceDao.selectRenamedScheme(scheme);// 判断该用户下是否存在同名体系的模板或实例
 		int num = -1;
-		if(list.size() == 0) {//若没有重名体系
+		if (list.size() == 0) {// 若没有重名体系
 			num = this.indiceDao.updateSchemeInfo(scheme);
 		}
 		response.setContentType("text/json;charset=utf-8");
@@ -351,9 +372,10 @@ public class IndiceController {
 		BufferedReader br = request.getReader();
 		String line = br.readLine();
 		IndiceInfo indice = JSON.parseObject(line, IndiceInfo.class);// 把JSON字符串解析为JavaBean
-		IndiceInfo fatherIndice = this.indiceDao.selectIndiceInfoByIndice_id(indice.getFather_id(), indice.getScheme_id());
+		IndiceInfo fatherIndice = this.indiceDao.selectIndiceInfoByIndice_id(indice.getFather_id(),
+				indice.getScheme_id());
 		int num = -1;
-		if(fatherIndice != null) {
+		if (fatherIndice != null) {
 			num = this.indiceDao.insertIndiceInfo(indice);
 		}
 		response.setContentType("text/json;charset=utf-8");
@@ -375,12 +397,13 @@ public class IndiceController {
 		String line = br.readLine();
 		IndiceInfo indice = JSON.parseObject(line, IndiceInfo.class);// 把JSON字符串解析为JavaBean
 		int num = -1;
-		if(indice.getFather_id() != -1) {//若不是根节点，需要判断父节点是否存在
-			IndiceInfo fatherIndice = this.indiceDao.selectIndiceInfoByIndice_id(indice.getFather_id(), indice.getScheme_id());
-			if(fatherIndice != null) {
+		if (indice.getFather_id() != -1) {// 若不是根节点，需要判断父节点是否存在
+			IndiceInfo fatherIndice = this.indiceDao.selectIndiceInfoByIndice_id(indice.getFather_id(),
+					indice.getScheme_id());
+			if (fatherIndice != null) {
 				num = this.indiceDao.updateIndiceInfo(indice);
 			}
-		}else {
+		} else {
 			num = this.indiceDao.updateIndiceInfo(indice);
 		}
 		response.setContentType("text/json;charset=utf-8");
@@ -516,21 +539,24 @@ public class IndiceController {
 			this.dfsForParseJSON(children.getJSONObject(i), indice.getIndice_id(), scheme_id);
 		}
 	}
+
 	/**
 	 * 获取服务器保存的计算结果
+	 * 
 	 * @param request
 	 * @param response
 	 * @throws FileNotFoundException
 	 * @throws IOException
 	 */
 	@RequestMapping(params = "request=downloadCalResult")
-	public void downloadCalResult(HttpServletRequest request, HttpServletResponse response) throws FileNotFoundException, IOException {
-		String filepath = Util.getCalcResultPath(request);//获取文件保存路径
+	public void downloadCalResult(HttpServletRequest request, HttpServletResponse response)
+			throws FileNotFoundException, IOException {
+		String filepath = Util.getCalcResultPath(request);// 获取文件保存路径
 		JSONObject currentUser = (JSONObject) (request.getSession().getAttribute("currentUser"));
 		UserInfo user = JSON.parseObject(currentUser.toString(), UserInfo.class);
 		int user_id = user.getUser_id();
-		int scheme_id = Integer.parseInt(request.getParameter("scheme_id"));//获取体系id
-		String scheme_name = this.indiceDao.selectScheme_nameByScheme_id(scheme_id);//根据体系id查询体系名称
+		int scheme_id = Integer.parseInt(request.getParameter("scheme_id"));// 获取体系id
+		String scheme_name = this.indiceDao.selectScheme_nameByScheme_id(scheme_id);// 根据体系id查询体系名称
 		String filename = scheme_name + "-" + user_id + ".xls";
 		// 保存在本地磁盘中的文件
 		File file = new File(filepath, filename);
@@ -548,7 +574,7 @@ public class IndiceController {
 		// 使用工具类直接将文件的字节复制到响应输出流中
 		FileCopyUtils.copy(new FileInputStream(file), response.getOutputStream());
 	}
-	
+
 	/**
 	 * 
 	 * @param request
@@ -560,17 +586,17 @@ public class IndiceController {
 	public void getAllResultTime(HttpServletResponse response, String scheme_id) throws IOException {
 		System.out.println("getTimeRequest");
 		List<Date> dateList = resultDao.selectTimeBySchemeId(Integer.parseInt(scheme_id));
-		
+
 		List<String> timeList = new ArrayList<>();
-		
+
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-		for(Date d: dateList) {
+		for (Date d : dateList) {
 			timeList.add(sdf.format(d));
 		}
-		
+
 		response.getWriter().write(JSON.toJSONString(timeList));
 	}
-	
+
 	/**
 	 * 
 	 * @param response
@@ -581,34 +607,36 @@ public class IndiceController {
 	 * @throws ParseException
 	 */
 	@RequestMapping(params = "request=getResult")
-	public void getAllResultTime(HttpServletResponse response, String scheme_id, String exec_time) throws IOException, NumberFormatException, ParseException {
+	public void getAllResultTime(HttpServletResponse response, String scheme_id, String exec_time)
+			throws IOException, NumberFormatException, ParseException {
 		System.out.println("getResultRequest");
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-		List<Result> resList = resultDao.selectResultBySchemeIdANDTime(Integer.parseInt(scheme_id), sdf.parse(exec_time));
-		
+		List<Result> resList = resultDao.selectResultBySchemeIdANDTime(Integer.parseInt(scheme_id),
+				sdf.parse(exec_time));
+
 		List<String> strResList = new ArrayList<>();
-		Map<Integer,List<Result>> map = new HashMap<>();
-		
-		for(Result res:resList) {
-			if(map.containsKey(res.getGroupId())) {
+		Map<Integer, List<Result>> map = new HashMap<>();
+
+		for (Result res : resList) {
+			if (map.containsKey(res.getGroupId())) {
 				map.get(res.getGroupId()).add(res);
-			}else {
+			} else {
 				List<Result> list = new ArrayList<>();
 				list.add(res);
 				map.put(res.getGroupId(), list);
 			}
 		}
-		
-		for(int i : map.keySet()) {
+
+		for (int i : map.keySet()) {
 			List<Result> tmpList = map.get(i);
-			
+
 			List<String> strList = new ArrayList<>();
-			for(Result r:tmpList) {
+			for (Result r : tmpList) {
 				strList.add(JSON.toJSONString(r));
 			}
 			strResList.add(JSON.toJSONString(strList));
 		}
-		
+
 		response.getWriter().write(JSON.toJSONString(strResList));
 	}
 }
